+++ packages/pic/src/pocket-ic-server.ts
@@ -94,7 +94,18 @@ export class PocketIcServer {
 
     return await poll(
       async () => {
-        const portString = await readFileAsString(portFilePath);
+        const portString = await readFileAsString(portFilePath).catch(
+          (err: unknown) => {
+            if (
+              err instanceof Error &&
+              (err as NodeJS.ErrnoException).code === 'ENOENT'
+            ) {
+              throw new BinTimeoutError();
+            }
+            throw err;
+          },
+        );
+
         const port = parseInt(portString);
         if (isNaN(port)) {
           throw new BinTimeoutError();
@@ -153,7 +164,7 @@ const POLL_INTERVAL_MS = 100;
 const POLL_TIMEOUT_MS = 90_000;
 
 class NullStream extends Writable {
-  _write(
+  override _write(
     _chunk: any,
     _encoding: BufferEncoding,
     callback: (error?: Error | null) => void,